+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+...
+
+## [0.2.0] - 2019-09-28
+
 ### Added
 - Method to read calibrated measurement of UVA, UVB and UV index.
 - Methods to set the integration time and dynamic setting.
@@ -20,5 +24,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 This is the initial release to crates.io. All changes will be documented in this CHANGELOG.
 
-[Unreleased]: https://github.com/eldruin/veml6075-rs/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/eldruin/veml6075-rs/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/eldruin/veml6075-rs/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/eldruin/veml6075-rs/releases/tag/v0.1.0